Koin modules split into different categories. Added AppInjector as an abstraction layour over DI. Minor code reformatting
