+++ app/src/main/java/com/akshay/newsapp/NewsApp.kt
@@ -1,20 +1,15 @@
 package com.akshay.newsapp
 
 import android.app.Application
-import com.akshay.newsapp.koin.appModule
-import org.koin.android.ext.koin.androidContext
-import org.koin.core.context.startKoin
+import com.akshay.newsapp.koin.AppInjector
 
 class NewsApp : Application() {
 
-    override fun onCreate() {
-        super.onCreate()
-        // start Koin!
-        startKoin {
-            // Android context
-            androidContext(applicationContext)
-            // modules
-            modules(appModule)
-        }
-    }
+	override fun onCreate() {
+		super.onCreate()
+
+		// Init DI magic ✨
+		AppInjector.init(this@NewsApp)
+
+	}
 }
