+++ app/src/main/java/com/akshay/newsapp/koin/modules/NewsDatabaseModule.kt
@@ -1,0 +1,23 @@
+package com.akshay.newsapp.koin.modules
+
+import com.akshay.newsapp.db.NewsDatabase
+import org.koin.android.ext.koin.androidApplication
+import org.koin.dsl.module
+
+/**
+ * All database related initializations
+ */
+
+val newsDatabaseModule = module {
+
+	// NewsDatabase Instance
+	single {
+		NewsDatabase.buildDefault(androidApplication())
+	}
+
+	// NewsArticleDao instance
+	single {
+		get<NewsDatabase>().newsArticlesDao()
+	}
+
+}
