+++ app/src/main/java/com/akshay/newsapp/koin/modules/RepositoryModule.kt
@@ -1,0 +1,16 @@
+package com.akshay.newsapp.koin.modules
+
+import com.akshay.newsapp.repo.NewsRepository
+import org.koin.dsl.module
+
+/**
+ * Repository initializations
+ */
+
+val repositoryModule = module {
+
+	// NewsRepository instance
+	single {
+		NewsRepository(get(), get())
+	}
+}
